backend/mlservice/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
pymongo==4.6.0
python-dotenv==1.0.0

